SBComponent -> ProviderContainer
